// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ShowTrackingStatus = {
  PLAN_TO_WATCH: 'PLAN_TO_WATCH',
  WATCHING: 'WATCHING',
  COMPLETED: 'COMPLETED',
  DROPPED: 'DROPPED'
} as const

export type ShowTrackingStatus = (typeof ShowTrackingStatus)[keyof typeof ShowTrackingStatus]


export const MovieTrackingStatus = {
  PLAN_TO_WATCH: 'PLAN_TO_WATCH',
  WATCHING: 'WATCHING',
  COMPLETED: 'COMPLETED'
} as const

export type MovieTrackingStatus = (typeof MovieTrackingStatus)[keyof typeof MovieTrackingStatus]